fixed serious bug [ 1757382 ] edit conflit undetected
Andre Engels did it properly in wikipedia.py CVS r1.927 (SVN r3710) with 
the description "Avoid 'solving' edit conflicts by throwing away the 
conflicting edit."

Bryan broke it in wikipedia.py CVS r1.950 (SVN r3837) with the 
description "Undo the change in r1.927 to get a token of the page itself 
instead of the sandbox, to prevent edit conflicts when editting as 
sysop."

I basically reverted it to Andre's version. I didn't test if unwanted 
edit conflicts are raised when someone edits the sandbox. But even if 
that's the case, having an edit conflict here and there is better than 
the catastrophic situation that we have at the moment (bots overwrite 
other user's changes).



git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk/pywikipedia@3899 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6