Follow up to Sébastien Rombants' bug #1981416 [Add the get_redirect argument to categories() & templates()] : * Adding get_redirect to Page::templates * Page::categories already had the nofollow_redirects argument : I tried searching what was that confusing nofollow_redirects argument :
It turns out that nofollow_redirects was a Page::get() parameter that was used *exactly* as get_redirect, except on two Page::_getEditPage() lines :
             if get_redirect:
                 self._redirarg = redirtarget
             elif not nofollow_redirects:
                 raise IsRedirectPage(redirtarget)
The only difference was that get_redirect was setting Page::_redirarg whereas nofollow_redirects wasn't.

Now, searching a bit more in our codebase, the only access to Page::_redirarg was in Page::get() :
             if hasattr(self, '_redirarg') and not get_redirect and not nofollow_redirects:
                 raise IsRedirectPage, self._redirarg
So, with nofollow_redirects=True, _redirarg wasnt set nor read, and IsRedirectPage was not ever raised.

Do you see a difference with the get_redirect=True behavior, where _redirarg is set but not read, and IsRedirectPage is not raised ? 

I don't.

=> Removing the duplicate nofollow_redirects, to use instead get_redirect



git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk/pywikipedia@5496 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6