Changing the GeneratorFactory so that instead of handleArg() being called once and returning a generator, it is called many times to accumulate a list of pages to work on, and then, at the end, the combined page generator is explicitly asked for.  This makes the GeneratorFactory a lot more useful for a wide variety of use cases.
Also, GeneratorFactory now understands the -page parameter, so some duplicate
code was removed from bots that handled this individually.

Not in this commit, but coming soon, will be the handling of -namespace in the
GeneratorFactory.  This will remove quite a bit of duplicate code from
individual bots and will make most bots more flexible.


git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk/pywikipedia@6270 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6