Bugfix: replace.py a b -xml:mydump.xml -namespace:0 also changed pages in other namespaces.
This is still not an ideal solution, as pages in other namespaces are still
processed in the XmlDumpReplacePageGenerator. This takes a lot of time,
especially for large deletion logs etc. which are in the Wikipedia
namespace.

cydeweys, russblau and the others: it would be nice if you would review my
changes, and also try to come up with a more elegant solution. Once we have
agreed on a good solution, we also need to make similar changes to other
scripts, such as selflink.py, table2wiki.py, and template.py. --Daniel




git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk/pywikipedia@6306 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6