Changed default number_of_words variable from 31 to 22.
The variable is utilized to limit number of words sent to search engine and influences the number of queries to execute for pages, as well as the number of results. The previous value was too low to work fine with Live Search, the current one is tested a long time on Italian Wikipedia.

git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk/pywikipedia@6598 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6